Worked on creating background services
Created a abstract NotificationService class, containing some base methods for sending Notifications, currently unsure of weather or not they actually work.

Created a abstract generic SignalRService class, that inherit from NotificationService. it is a more specified NotificationService, focused around processing messages from SignalR.

Created a extensions class, to simplify the the process of starting a service. i'm unsure why the extensions it have made, deos not exist already, as it is very small and simple.