Fixed Notifications not firing. Was missing a few things, including but not limited to:
- The Id of the icon that should be displayed with the notification.
- The notification priority
- Was using Notification.Builder, when i was suppose to use NotificationCompat.Builder

Updated CreateNotificationChannel to take in the text for the description, as it made no sense that it was constantly saying that it was caused by a SignalR Hub, as the NotificationService class can be used for other services.

Added IsStarted boolean, that is set to true during OnCreate, and set to false prior to stopping the service. That piece of code is further up the inheritence, but should proably be done in the NotificationService class.

Added logging to most if not all important places in NotificationService. Also includes Try/Catch, with Logging of any exceptions that get caught, and just re-throwing the exceptions.